requirements.txt
Django>=4.0.0,<4.2.0
djangorestframework>=3.12.0,<3.14.0
djangorestframework-simplejwt>=5.0.0,<5.3.0
drf-yasg>=1.20.0,<1.21.7
django-cors-headers>=3.10.0,<4.3.1
Pillow>=9.0.0,<10.2.0
python-dotenv>=0.19.0,<1.0.0